Too Much
I spent a while without committing and live to regret my choices

Bigger things
- Framebuffer device improvements
- VGA mode setting for framebuffer devices
- fmove now does a search for the next greater slot to insert into a
  file table (more POSIX-y)

+++ include/kanawha/uapi/process.h
@@ -1,8 +1,6 @@
 #ifndef __KANAWHA__UAPI_PROCESS_H__
 #define __KANAWHA__UAPI_PROCESS_H__
 
-#include <kanawha/uapi/stdint.h>
-
-typedef long pid_t;
+typedef int pid_t;
 
 #endif
